Move workflow tests to own module, update with changed requirements.
